refactor: keep recorder meter local
